* Change the way the user is viewed (#431)
