+++ frontend/src/components/Navbar/Navbarcopy.module.css
@@ -7,9 +7,70 @@
 .brandText {
     font-size: 1.4rem;
     font-weight: 500;
-    white-space: nowrap;   
+    white-space: nowrap;
+    }
+
+
+.profileButton {
+    position: relative;
 }
 
+.profileButton:hover .profileCircle .tooltip {
+    opacity: 1;
+    visibility: visible;
+}
+
+.profileButton.dropdownOpen .profileCircle .tooltip {
+    opacity: 0 !important;
+    visibility: hidden !important;
+}
+
+.profileWrapper {
+    position: relative;
+}
+
+/* Tooltip */
+.tooltip {
+    position: absolute;
+    top: calc(100% + 8px);
+    left: 50%;
+    transform: translateX(-50%);
+    padding: 8px 12px;
+    background-color: #1f2937;
+    color: white;
+    font-size: 14px;
+    border-radius: 8px;
+    box-shadow: 0 10px 15px -3px rgba(0, 0, 0, 0.1), 0 4px 6px -2px rgba(0, 0, 0, 0.05);
+    opacity: 0;
+    visibility: hidden;
+    transition: opacity 0.2s ease-in-out;
+    pointer-events: none;
+    white-space: nowrap;
+    z-index: 50;
+}
+
+.tooltipContent {
+    display: flex;
+    flex-direction: column;
+}
+
+.userRole {
+    font-size: 12px;
+    color: #d1d5db;
+}
+
+
+.tooltipArrow {
+    position: absolute;
+    bottom: 100%;
+    left: 50%;
+    transform: translateX(-50%);
+    width: 0;
+    height: 0;
+    border-left: 4px solid transparent;
+    border-right: 4px solid transparent;
+    border-bottom: 4px solid #1f2937;
+}
 .brandText2 {
     font-size: 1rem;
     color: #4a5565;
@@ -26,7 +87,6 @@
     border-radius: 2px;
     vertical-align: middle;
 }
-
 /* Profile Card Styles */
 .profileCard {
     display: flex;
@@ -45,20 +105,30 @@
 }
 
 .profileCircle {
-    background-color: #e6f9ed;
-    color: #157347;
-    border-radius: 50%;
     width: 32px;
     height: 32px;
+    background-color: #dcfce7;
+    border-radius: 50%;
     display: flex;
     align-items: center;
     justify-content: center;
-    font-weight: bold;
+    color: #166534;
+    font-weight: 500;
+    position: relative;
+    border: 2px solid;
+}
+
+.adminBorder {
+    border-color: #15803d;
+}
+
+.userBorder {
+    border-color: #d1d5db;
 }
 
 .userName {
     font-weight: 500;
-    color: #333;
+    margin-bottom: 2px;
 }
 
 .financialToggleText {
@@ -66,7 +136,7 @@
     color: #333;
 }
 
-.financialToggle{
+.financialToggle {
     cursor: pointer;
     height: 1.25rem;
     width: 2.25rem;
@@ -80,7 +150,7 @@
 }
 
 .financialToggle[type="checkbox"]:checked {
-    background-color: #43A047; 
+    background-color: #43A047;
     border-color: #43A047;
 }
 
@@ -90,7 +160,7 @@
 }
 
 @media (max-width: 900px) {
-    .financialToggleText{
+    .financialToggleText {
         display: none;
     }
 }
@@ -114,8 +184,8 @@
 }
 
 .headerNavbar {
-    -ms-flex-wrap: nowrap !important; 
-    flex-wrap: nowrap  !important;
+    -ms-flex-wrap: nowrap !important;
+    flex-wrap: nowrap !important;
     display: flex;
     justify-content: center;
     align-items: center;
@@ -124,24 +194,19 @@
 
 @media (max-width: 991.98px) {
     .headerNavbar {
-        -ms-flex-wrap: nowrap !important; 
-        flex-wrap: nowrap  !important;
+        -ms-flex-wrap: nowrap !important;
+        flex-wrap: nowrap !important;
         display: flex;
         justify-content: center;
         align-items: center;
         gap: 15px;
     }
-    .headerNavbarAlt{
-        -ms-flex-wrap: nowrap !important; 
-        flex-wrap: nowrap  !important;
+
+    .headerNavbarAlt {
+        -ms-flex-wrap: nowrap !important;
+        flex-wrap: nowrap !important;
         display: flex;
         gap: 68%;
-    }
-}
-
-@media (max-width: 880px){
-    .userName {
-    display: none !important; 
     }
 }
 
@@ -151,7 +216,7 @@
     gap: 1rem;
 }
 
-.appHeader{
+.appHeader {
     position: fixed;
     z-index: 10;
     width: 100%;
